[#69481358 #69481360 #69481362] User can register, login, and logout
also updated views to match mockups, user can refresh index page without form resend confirmation message. Please note that tests don't accurately prove that error message is displayed for invalid password, but the code is included in the application at this point. We'll include this test in the appropriate story

Signed-off-by: Ellie Schneiders <ems2141@gmail.com>